Modifying the coefficient type of terms to double. Added necessary LP model extension methods for the first phase of the two-phase simplex. Added the skeleton of the two-phase simplex algoritm.
Added SimplexAlgoritmExecutionException for signing the errors occured during the execution of the simplex algoritm. Editing the comments, minor cosmethic refactorings in the code.

+++ simplexapi/simplexapi.Common/Extensions/EquationExtensions.cs
@@ -25,7 +25,7 @@ namespace simplexapi.Common.Extensions
             return equation;
         }
 
-        public static Equation Multiply(this Equation equation, int factor)
+        public static Equation Multiply(this Equation equation, double factor)
         {
             equation.LeftSide.ForAll(term => term.SignedCoefficient *= factor);
             equation.RightSide.ForAll(term => term.SignedCoefficient *= factor);
